requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.2.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2025.3.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
greenlet==3.0.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib-metadata==4.6.4
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.92
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
narwhals==1.32.0
neo4j==5.14.1
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai==0.27.10
openpyxl==3.1.5
orjson==3.10.6
packaging==23.2
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
plotly==5.18.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==4.25.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.2.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.0
SecretStorage==3.3.1
sentence-transformers==3.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.29.0
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
texttable==1.6.4
threadpoolctl==3.5.0
tiktoken==0.3.3
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.2.3
websockets==14.2
Werkzeug==3.1.3
XlsxWriter==3.2.0
yarl==1.9.4
zipp==1.0.0
